JSONB keys are less useful than hoped
We still can't use JSON objects as dict keys (which we need for
channel identifiers).

Also, sqlalchemy appears to have some weird behavior where if you pass
a simple string in for a JSON value, it doesn't JSON-encode it? Maybe
it assumes that JSON types are always "objects", and if you pass a
string then it means you have some pre-encoded JSON? I can't find this
in the source, but I don't know how else to explain the errors I'm
seeing.